Made some progress

+++ db/migrate/20191010020952_add_player.rb
@@ -2,8 +2,8 @@ class AddPlayer < ActiveRecord::Migration[6.0]
   def change
     create_table :players do |t|
       t.string :name
-      t.integer :ratio
-      t.integer :luck
+      t.decimal :ratio
+      t.decimal :luck
       t.integer :wins
       t.integer :losses
       t.integer :luckwins
